tasks/web/pretty/src/requirements.txt
autopep8==2.0.4
beautifulsoup4==4.12.2
Flask==2.2.4
lxml==4.9.3
peewee==3.16.2
pydantic==1.10.2
PyJWT==2.8.0
gunicorn==21.2.0